refactor(components/atom/videoPlayer): Reduce vdom tree reconciliation load by replacing the switche
